refactor: separate UserInterface class
UserInterface 클래스를 InputView와 OuputView로 분리했습니다. InputView는
사용자 입력을 받고 처리하며, OutputView는 사용자에게 보여주는 출력을
관리합니다.
