+++ src/test/java/racingcar/ApplicationTest.java
@@ -7,20 +7,6 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class ApplicationTest extends NsTest {
-//    private static final int MOVING_FORWARD = 4;
-//    private static final int STOP = 3;
-//
-//    @Test
-//    void 기능_테스트() {
-//        assertRandomNumberInRangeTest(
-//            () -> {
-//                run("pobi,woni", "1");
-//                assertThat(output()).contains("pobi : -", "woni : ", "최종 우승자 : pobi");
-//            },
-//            MOVING_FORWARD, STOP
-//        );
-//    }
-
     @Test
     void 예외_테스트() {
         assertSimpleTest(() ->
