+++ src/test/java/racingcar/view/OutputViewTest.java
@@ -9,20 +9,20 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import racingcar.model.Car;
 
-class UserInterfaceTest {
+class OutputViewTest {
 
     @Test
     void 단독우승_출력_성공() {
         // given
         Car car1 = new Car("car1");
         List<Car> winners = List.of(car1);
-        UserInterface userInterface = new UserInterface();
+        OutputView outputView = new OutputView();
 
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         System.setOut(new PrintStream(outputStream));
 
         // when
-        userInterface.showFinalWinner(winners);
+        outputView.showFinalWinner(winners);
 
         // then
         String expectedOutput = "최종 우승자 : " + "car1" + System.lineSeparator();
@@ -36,14 +36,14 @@ class UserInterfaceTest {
         Car car1 = new Car("car1");
         Car car2 = new Car("car2");
         Car car3 = new Car("car3");
+        OutputView outputView = new OutputView();
         List<Car> winners = Arrays.asList(car1, car2, car3);
-        UserInterface userInterface = new UserInterface();
 
         ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
         System.setOut(new PrintStream(outputStream));
 
         // when
-        userInterface.showFinalWinner(winners);
+        outputView.showFinalWinner(winners);
 
         // then
         String expectedOutput = "최종 우승자 : " + "car1, car2, car3" + System.lineSeparator();
